Converted main to Grammar Test Loop
Adds initial files for gradual conversion and testing of the new grammar specification
